refactor http code, add protobuf runtime
