+++ lib/distribution.js
@@ -1,7 +1,7 @@
 const Hyperswarm = require('hyperswarm')
 const crypto = require('crypto')
 const {DiscoverySwarm} = require('./discovery/swarm')
-const {handleRequests} = require('./http-utils')
+const {handleRequests} = require('./http/utils')
 const debug = require('debug')('fm2u-distribution')
 
 const swarms = new Map()
