+++ lib/http/utils.js
@@ -1,6 +1,6 @@
 const pathMatch = require('path-match')
 const {parse: parseUrl} = require('url')
-const {parseRequest} = require('./http-request')
+const {parseRequest} = require('./request')
 
 const routeMatcher = pathMatch({
   sensitive: true,
